Round the interpreter's multiply before its add
Go may fuse x*y + z into one FMA across statements, and does on arm64
and at GOAMD64=v3, which opMulAddRKR's float path let it do. C Lua
rounds the product, as compiled code does; an explicit conversion stops
the fusion (Go spec, Arithmetic operators).

+++ lua/vm_jit.go
@@ -233,7 +233,7 @@ func (l *State) executeSwitchJIT() {
 			}
 		case opMulAddRKR:
 			if b, c := frame[i.B()], constants[i.C()]; b.isFloat() && c.isFloat() {
-				r := b.f() * c.f()
+				r := float64(b.f() * c.f()) // rounded: Go may fuse x*y + z otherwise, as C Lua does not
 				frame[i.A()] = numberValue(r)
 				if j := code[ip]; l.hookMask&(MaskLine|MaskCount) == 0 {
 					if d := frame[j.C()]; d.isFloat() {
